chore(tui): split oversized command.rs into command/ module
Extract 1778-line src/tui/command.rs into four sub-modules:

- command/mod.rs (17 lines): module declarations + re-exports

- command/specs.rs (302 lines): COMMAND_SPECS, parsing, matching, help text

- command/status.rs (832 lines): status/context rendering, model help, download status

- command/tests.rs (517 lines): all tests

Also fix pre-existing type errors: COMMAND_SPECS count (18->19),

LocalCommandKind::ModelName->Model, has_active_thread() -> inline check,

cache_dir/state_db_path -> state_db_status.
